# swagger.yaml
openapi: 3.0.0
info:
  title: Node.js Sign-up and Verification API
  description: Node.js + MySQL - API with email sign-up, verification, authentication and forgot password
  version: 1.0.0

servers:
  - url: http://localhost:4000
    description: Local development server
  - url: https://ipt-2026-backend-xfoa.onrender.com
    description: Production server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      in: header
      name: Authorization

paths:
  /accounts/authenticate:
    post:
      summary: Authenticate account credentials and return a JWT token and a cookie with a refresh token
      operationId: authenticate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: Dianne@example.com
                password:
                  type: string
                  example: Dianne2004
      responses:
        "200":
          description: OK

  /accounts/refresh-token:
    post:
      summary: Use a refresh token to generate a new JWT token and a new refresh token
      operationId: refreshToken
      responses:
        "200":
          description: OK

  /accounts/revoke-token:
    post:
      summary: Revoke a refresh token
      operationId: revokeToken
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
      responses:
        "200":
          description: OK

  /accounts/register:
    post:
      summary: Register a new user account and send a verification email
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: Ms
                firstName:
                  type: string
                  example: Dianne
                lastName:
                  type: string
                  example: Escuadro
                email:
                  type: string
                  example: Dianne@example.com
                password:
                  type: string
                  example: Dianne2004
                confirmPassword:
                  type: string
                  example: Dianne2004
                acceptTerms:
                  type: boolean
                  example: true
      responses:
        "200":
          description: OK

  /accounts/verify-email:
    post:
      summary: Verify a new account with a verification token received by email after registration
      operationId: verifyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
      responses:
        "200":
          description: OK

  /accounts/forgot-password:
    post:
      summary: Submit email address to reset the password on an account
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: Dianne@example.com
      responses:
        "200":
          description: OK

  /accounts/validate-reset-token:
    post:
      summary: Validate the reset password token received by email
      operationId: validateResetToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
      responses:
        "200":
          description: OK

  /accounts/reset-password:
    post:
      summary: Reset the password for an account
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                password:
                  type: string
                  example: newPassword2004
                confirmPassword:
                  type: string
                  example: newPassword2004
      responses:
        "200":
          description: OK

  /accounts:
    get:
      summary: Get a list of all accounts (Admin only)
      operationId: getAll
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
    post:
      summary: Create a new account (Admin only)
      operationId: create
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: Ms
                firstName:
                  type: string
                  example: Jane
                lastName:
                  type: string
                  example: Doe
                email:
                  type: string
                  example: Jane@example.com
                password:
                  type: string
                  example: jane123
                confirmPassword:
                  type: string
                  example: jane123
                role:
                  type: string
                  example: Admin
      responses:
        "200":
          description: OK

  /accounts/{id}:
    get:
      summary: Get a single account by ID
      operationId: getById
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
    put:
      summary: Update an account
      operationId: update
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
                password:
                  type: string
                confirmPassword:
                  type: string
                role:
                  type: string
      responses:
        "200":
          description: OK
    delete:
      summary: Delete an account
      operationId: delete
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK